Improve serve test, add single txt file test
