+++ rojo-test/serve-tests/just-txt.txt
@@ -1,0 +1,1 @@
+Hello, world!
