+++ rojo-test/src/lib.rs
@@ -2,4 +2,5 @@
 
 mod build_test;
 mod serve_test;
+mod serve_util;
 mod util;
